% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovoExample.R
\name{cleanPopulation_ovo}
\alias{cleanPopulation_ovo}
\title{cleanPopulation}
\usage{
cleanPopulation_ovo(pop, clf)
}
\arguments{
\item{pop:}{is population (a list) of predomics objects}

\item{clf:}{the classifier object}
}
\value{
a list population of predomics objects
}
\description{
Looks for invalid predomics objects in a population and removes them.
}
