% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovoExample.R
\name{computeConfusionMatrix_ovo}
\alias{computeConfusionMatrix_ovo}
\title{Evaluates the confusion Matrix of the predicted class and the class to predict}
\usage{
computeConfusionMatrix_ovo(mod, X, y, clf)
}
\arguments{
\item{mod:}{a model object to be evaluated}

\item{X:}{dataset to classify}

\item{y:}{variable to predict}

\item{clf:}{an object containing the different parameters of the classifier}
}
\value{
a list confusion matrix
}
\description{
This function evaluates the accuracy of a model
}
