% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovoExample.R
\name{evaluateAccuracy_ovo}
\alias{evaluateAccuracy_ovo}
\title{Evaluates the accuracy of a model}
\usage{
evaluateAccuracy_ovo(
  mod = NULL,
  X,
  y,
  clf,
  force.re.evaluation = FALSE,
  mode = "train"
)
}
\arguments{
\item{mod:}{a model object to be used in the class prediction}

\item{X:}{dataset to classify}

\item{y:}{variable to predict}

\item{clf:}{an object containing the different parameters of the classifier}

\item{force.re.evaluation:}{evaluate again all the elements needed for accuracy (default:FALSE)}

\item{mode:}{training or test mode. If training, the funciton maximizes accuracy.}
}
\value{
either (1) list a model whose evaluation parameters are updated or (2) the accuracy
}
\description{
This function evaluates the accuracy of either (1) a model object that contains intercept and sign or (2) directly the attributes score, intercept, sign.one versus one
}
