% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalova.R
\name{evaluateAdditionnalMetrics_ova}
\alias{evaluateAdditionnalMetrics_ova}
\title{One Versus All, Compute other prediction scores such as precision, recall and f-score a model}
\usage{
evaluateAdditionnalMetrics_ova(mod, X, y, clf, mode = "train")
}
\arguments{
\item{mod:}{a model object to be evaluated}

\item{X:}{dataset to classify}

\item{y:}{variable to predict}

\item{clf:}{an object containing the different parameters of the classifier}

\item{mode:}{training or testing mode}
}
\value{
a model whose evaluation parameters are updated
}
\description{
One Versus All, This function computes prediction scores based on the confusion matrix such as accuracy, precision, recall and f-score a model one versus all
}
