% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R, R/ovoExample.R
\name{evaluateAUC_ovo}
\alias{evaluateAUC_ovo}
\title{Computes the AUC of a model}
\usage{
evaluateAUC_ovo(score, y, sign = ">")

evaluateAUC_ovo(score, y, sign = ">")
}
\arguments{
\item{score:}{the ^y score of the model}

\item{y:}{the response vector}

\item{sign:}{in which direction to make the comparison? "auto" (default): automatically define in which group
the median is higher and take the direction accordingly. ">": if the predictor values for the control group
are higher than the values of the case group (controls > t >= cases). "<": if the predictor values for the
control group are lower or equal than the values of the case group (controls < t <= cases).}
}
\value{
a list auc value

an auc value
}
\description{
One Versus one, Computes the AUC of a model

Computes the AUC of a model
}
