% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovoExample.R
\name{evaluateFit_ovo}
\alias{evaluateFit_ovo}
\title{evaluateFit_ovo}
\usage{
evaluateFit_ovo(mod, X, y, clf, force.re.evaluation = FALSE, mode = "train")
}
\arguments{
\item{mod}{: a model object}

\item{X:}{the data matrix with variables in the rows and observations in the columns}

\item{y:}{the response vector}

\item{clf:}{the classifier parameter object}

\item{force.re.evaluation:}{re-evaluate all the scores even if they exist (default:FALSE)}

\item{mode:}{A choice from c("train", "test") indicates wether we wish to learn the threthold
of the model (default:"train") or not "test" for the c("terinter","bininter","ratio") languages}
}
\value{
a model object with the fitting score
Evaluates the fitting score of a model object

a list of model objects with the appropriate score
}
\description{
Evaluates the fitting score of a model object.

Evaluates the fitting score of a model object.
}
\details{
Evaluates the fitting score of a model object
}
