% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalova.R
\name{evaluateIntercept_ova}
\alias{evaluateIntercept_ova}
\title{Evaluates the fitting score of a model object}
\usage{
evaluateIntercept_ova(mod, X, y, clf)
}
\arguments{
\item{mod}{: a model object}

\item{X:}{the data matrix with variables in the rows and observations in the columns}

\item{y:}{the response vector}

\item{clf:}{the classifier parameter object}
}
\value{
a list model object with the fitting score
}
\description{
Evaluates the fitting score of a model object.
}
