% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovoExample.R
\name{evaluateIntercept_ovo}
\alias{evaluateIntercept_ovo}
\title{Evaluates the fitting scores of a model objects}
\usage{
evaluateIntercept_ovo(mod, X, y, clf)
}
\arguments{
\item{mod}{: model objects}

\item{X:}{the data matrix with variables in the rows and observations in the columns}

\item{y:}{the response vector}

\item{clf:}{the classifier parameter object}
}
\value{
model objects with fit scores
}
\description{
Evaluates fit scores of model objects.
}
