% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovoExample.R
\name{getModelScore_ovo}
\alias{getModelScore_ovo}
\title{Computes the ^y score of the model}
\usage{
getModelScore_ovo(mod, X, clf, force.re.evaluation = TRUE)
}
\arguments{
\item{mod:}{a model object where the score will be computed}

\item{X:}{the data matrix with variables in the rows and observations in the columns}

\item{force.re.evaluation:}{we recompute the score (default:TRUE)}
}
\value{
a list containing the predicted ^y score for each observation
}
\description{
Returns the ^y score of the model
}
