% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R, R/ovoExample.R
\name{getSign_ovo}
\alias{getSign_ovo}
\title{Evaluates the sign for a given feature this is the old getMgsVsTraitSignDiscr function}
\usage{
getSign_ovo(X, y, clf = NULL, parallel.local = FALSE)

getSign_ovo(X, y, clf = NULL, parallel.local = FALSE)
}
\arguments{
\item{X:}{the data matrix with variables in the rows and observations in the columns}

\item{y:}{the response vector}

\item{clf:}{the classifier parameter object}

\item{parallel.local:}{weather or not to run in //}
}
\value{
a list of +1 & -1 for each one-vs-all classification of our dataset

a list of +1 & -1 for each one-vs-all classification of our dataset
}
\description{
Evaluates the sign for a given feature this is the old getMgsVsTraitSignDiscr function, one versus one.

Evaluates the sign for a given feature this is the old getMgsVsTraitSignDiscr function, one versus one.
}
